pyproject.toml
[project]
name = "async-relay"
version = "1.1.0"
description = "A high-performance TCP/UDP relay service with automatic failover"
authors = [
    {name = "So0ni"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["relay", "proxy", "tcp", "udp", "failover", "asyncio", "uvloop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: System :: Networking",
]

dependencies = [
    "pyyaml>=6.0.1",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "watchdog>=3.0.0",
    "aiohttp>=3.9.0",
    "aiohttp-basicauth>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
relay = "src.cli:run"

[project.urls]
Homepage = "https://github.com/So0ni/async-relay"
Repository = "https://github.com/So0ni/async-relay"
Issues = "https://github.com/So0ni/async-relay/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
include = ["src/static/**"]

[tool.hatch.build.targets.sdist]
include = ["src/static/**"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "watchdog.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
